New: "Check CreateAsync Result HttpPost Register Action"
Check the Result property of the HttpPost Register's CreateAsync call if Result.Succeeded.

If Result.Succeeded is false, foreach error in Result.Errors use ModelState.AddModelError to add an error with a the first parameter of"Password" and a second with the value of the error's Description property.

Then return the Register view with the model provided by Register's the parameter.